+++ src/tasks.rs
@@ -23,20 +23,35 @@ use crate::{input, time, todoist};
 #[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
 pub struct Task {
     pub id: String,
-    pub content: String,
-    pub priority: Priority,
-    pub description: String,
-    pub labels: Vec<String>,
-    pub parent_id: Option<String>,
+    pub user_id: String,
     pub project_id: String,
-    pub comment_count: Option<u32>,
-    pub due: Option<DateInfo>,
-    /// Only on rest api return value
-    pub is_completed: Option<bool>,
-    pub is_deleted: Option<bool>,
-    /// only on sync api return value
-    pub checked: Option<bool>,
+    pub section_id: Option<String>,
+    pub parent_id: Option<String>,
+    pub added_by_uid: Option<String>,
+    pub assigned_by_uid: Option<String>,
+    pub responsible_uid: Option<String>,
+    pub labels: Vec<String>,
+    pub deadline: Option<Deadline>,
     pub duration: Option<Duration>,
+    pub due: Option<DateInfo>,
+    pub checked: bool,
+    pub is_deleted: bool,
+    pub is_collapsed: bool,
+    pub added_at: Option<String>,
+    pub completed_at: Option<String>,
+    pub updated_at: Option<String>,
+    pub priority: Priority,
+    pub child_order: i16,
+    pub content: String,
+    pub description: String,
+    pub note_count: u32,
+    pub day_order: i16,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
+pub struct TaskResponse {
+    pub results: Vec<Task>,
+    pub next_cursor: Option<String>,
 }
 
 // Update task_attributes fn when adding here
@@ -88,10 +103,23 @@ impl Display for Task {
 
 #[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
 pub struct DateInfo {
+    /// "2025-04-26T22:00:00Z"
     pub date: String,
     pub is_recurring: bool,
+    /// "2025-04-26 15:00"
     pub string: String,
+    /// i.e. "en"
+    pub lang: String,
+    /// i.e. "America/Vancouver"
     pub timezone: Option<String>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
+pub struct Deadline {
+    /// In format YYYY-MM-DD
+    pub date: String,
+    /// i.e. "en"
+    pub lang: String,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
@@ -195,7 +223,7 @@ impl Task {
             format::labels(self)
         };
 
-        let comment_count = self.comment_count.unwrap_or_default();
+        let comment_count = self.note_count;
 
         let comment_number = match comment_count {
             0 => String::new(),
@@ -753,19 +781,13 @@ pub fn spawn_update_task_priority(
     })
 }
 
-pub fn sync_json_to_tasks(json: String) -> Result<Vec<Task>, Error> {
-    let body: Body = serde_json::from_str(&json)?;
-    Ok(body.items)
-}
-
-pub fn rest_json_to_tasks(json: String) -> Result<Vec<Task>, Error> {
-    let tasks: Vec<Task> = serde_json::from_str(&json)?;
-    Ok(tasks)
-}
-
 pub fn json_to_task(json: String) -> Result<Task, Error> {
     let task: Task = serde_json::from_str(&json)?;
     Ok(task)
+}
+pub fn json_to_tasks(json: String) -> Result<Vec<Task>, Error> {
+    let response: TaskResponse = serde_json::from_str(&json)?;
+    Ok(response.results)
 }
 
 pub fn sort_by_value(mut tasks: Vec<Task>, config: &Config) -> Vec<Task> {
@@ -798,7 +820,7 @@ pub async fn reject_parent_tasks(tasks: Vec<Task>, config: &Config) -> Vec<Task>
     let parent_ids: Vec<String> = tasks
         .clone()
         .into_iter()
-        .filter(|task| task.parent_id.is_some() && !task.checked.unwrap_or_default())
+        .filter(|task| task.parent_id.is_some() && !task.checked)
         .map(|task| task.parent_id.unwrap_or_default())
         .collect();
 
@@ -812,7 +834,7 @@ pub async fn reject_parent_tasks(tasks: Vec<Task>, config: &Config) -> Vec<Task>
         let config = config.clone();
         let handle = tokio::spawn(async move {
             if !parent_ids.contains(&task.id)
-                && !task.checked.unwrap_or_default()
+                && !task.checked
                 && !parent_in_future(task.clone(), tasks, &config).await
             {
                 Some(task)
@@ -848,7 +870,7 @@ async fn parent_in_future(task: Task, tasks: Vec<Task>, config: &Config) -> bool
                 false
             } else {
                 // look up id and see if it is in the future
-                match todoist::get_task(config, ID::Legacy(parent_id)).await {
+                match todoist::get_task(config, &parent_id).await {
                     Err(e) => {
                         config.clone().tx().send(e).unwrap();
                         false
@@ -917,6 +939,7 @@ mod tests {
             due: Some(DateInfo {
                 date: String::from("2001-11-13"),
                 is_recurring: true,
+                lang: String::from("en"),
                 timezone: Some(String::from("America/Los_Angeles")),
                 string: String::from("Every 2 weeks"),
             }),
@@ -1045,6 +1068,7 @@ mod tests {
         let task_today = Task {
             due: Some(DateInfo {
                 date: time::today_string(&config).unwrap(),
+                lang: String::from("en"),
                 is_recurring: false,
                 string: String::from("Every 2 weeks"),
                 timezone: None,
@@ -1057,6 +1081,7 @@ mod tests {
             due: Some(DateInfo {
                 date: String::from("2021-09-06T16:00:00"),
                 is_recurring: false,
+                lang: String::from("en"),
                 timezone: None,
                 string: String::from("Every 2 weeks"),
             }),
@@ -1071,6 +1096,7 @@ mod tests {
         let today = Task {
             due: Some(DateInfo {
                 date: time::today_string(&config).unwrap(),
+                lang: String::from("en"),
                 is_recurring: false,
                 timezone: None,
                 string: String::from("Every 2 weeks"),
@@ -1082,6 +1108,7 @@ mod tests {
             due: Some(DateInfo {
                 date: time::today_string(&config).unwrap(),
                 is_recurring: false,
+                lang: String::from("en"),
                 string: String::from("Every 2 weeks"),
                 timezone: None,
             }),
@@ -1092,6 +1119,7 @@ mod tests {
             due: Some(DateInfo {
                 date: String::from("2035-12-12"),
                 is_recurring: false,
+                lang: String::from("en"),
                 string: String::from("Every 2 weeks"),
                 timezone: None,
             }),
@@ -1109,9 +1137,21 @@ mod tests {
         let config = test::fixtures::config().await;
         let no_date = Task {
             id: String::from("222"),
+            section_id: None,
+            user_id: String::from("222"),
             content: String::from("Get gifts for the twins"),
-            checked: None,
-            comment_count: None,
+            checked: false,
+            child_order: 0,
+            day_order: 0,
+            updated_at: None,
+            deadline: None,
+            completed_at: None,
+            added_at: None,
+            added_by_uid: None,
+            responsible_uid: None,
+            assigned_by_uid: None,
+            note_count: 0,
+            is_collapsed: false,
             parent_id: None,
             project_id: String::from("123"),
             description: String::from(""),
@@ -1122,14 +1162,14 @@ mod tests {
             due: None,
             labels: vec![String::from("computer")],
             priority: Priority::Medium,
-            is_deleted: None,
-            is_completed: None,
+            is_deleted: false,
         };
 
         let date_not_datetime = Task {
             due: Some(DateInfo {
                 date: time::today_string(&config).unwrap(),
                 is_recurring: false,
+                lang: String::from("en"),
                 string: String::from("Every 2 weeks"),
                 timezone: None,
             }),
@@ -1140,6 +1180,7 @@ mod tests {
             due: Some(DateInfo {
                 date: String::from("2020-09-06T16:00:00"),
                 is_recurring: false,
+                lang: String::from("en"),
                 string: String::from("Every 2 weeks"),
                 timezone: None,
             }),
@@ -1150,6 +1191,7 @@ mod tests {
             due: Some(DateInfo {
                 date: String::from("2035-09-06T16:00:00"),
                 string: String::from("Every 2 weeks"),
+                lang: String::from("en"),
                 is_recurring: false,
                 timezone: None,
             }),
@@ -1159,6 +1201,7 @@ mod tests {
         let past = Task {
             due: Some(DateInfo {
                 date: String::from("2015-09-06T16:00:00"),
+                lang: String::from("en"),
                 is_recurring: false,
                 string: String::from("Every 2 weeks"),
                 timezone: None,
@@ -1183,18 +1226,29 @@ mod tests {
         let config = test::fixtures::config().await;
         let task = Task {
             id: String::from("222"),
-            content: String::from("Get gifts for the twins"),
-            checked: None,
+            section_id: None,
+            added_by_uid: None,
+            responsible_uid: None,
+            assigned_by_uid: None,
+            added_at: None,
+            is_collapsed: false,
+            user_id: String::from("222"),
+            checked: false,
+            child_order: 0,
+            day_order: 0,
+            deadline: None,
+            updated_at: None,
             duration: None,
+            completed_at: None,
             parent_id: None,
-            comment_count: Some(1),
+            note_count: 1,
+            content: String::from("Get gifts for the twins"),
             description: String::from(""),
             project_id: String::from("123"),
             labels: vec![String::from("computer")],
             due: None,
             priority: Priority::Medium,
-            is_deleted: None,
-            is_completed: None,
+            is_deleted: false,
         };
 
         assert!(!task.is_overdue(&config).unwrap());
@@ -1202,6 +1256,7 @@ mod tests {
         let task_today = Task {
             due: Some(DateInfo {
                 date: time::today_string(&config).unwrap(),
+                lang: String::from("en"),
                 string: String::from("Every 2 weeks"),
                 is_recurring: false,
                 timezone: None,
@@ -1213,6 +1268,7 @@ mod tests {
         let task_future = Task {
             due: Some(DateInfo {
                 date: String::from("2035-12-12"),
+                lang: String::from("en"),
                 is_recurring: false,
                 string: String::from("Every 2 weeks"),
                 timezone: None,
@@ -1224,6 +1280,7 @@ mod tests {
         let task_today = Task {
             due: Some(DateInfo {
                 date: String::from("2020-12-20"),
+                lang: String::from("en"),
                 is_recurring: false,
                 string: String::from("Every 2 weeks"),
                 timezone: None,
@@ -1246,7 +1303,7 @@ mod tests {
         let task = test::fixtures::task();
         let mut server = mockito::Server::new_async().await;
         let mock = server
-            .mock("POST", "/rest/v2/tasks/222")
+            .mock("POST", "/api/v1/tasks/6Xqhv4cwxgjwG9w8")
             .with_status(200)
             .with_header("content-type", "application/json")
             .with_body(test::responses::task())
@@ -1296,7 +1353,7 @@ mod tests {
     #[tokio::test]
     async fn test_display_task() {
         let task = test::fixtures::task();
-        let string = String::from("Get gifts for the twins");
+        let string = String::from("TEST");
         assert_eq!(string, task.to_string())
     }
 }